Support Multiple Languages at the Same Time
This update adds support for multiple languages at the same time.
This feature is still in development and may not work perfectly for your purposes.
It works by asking google for Speech-to-Text for the audio in two languages,
next it uses google translate to translate the text to English if in another
language. This requires the goslate python package which can be installed
with: ```sudo pip install goslate```