FIX: Changed to Google Speech API V2
Google has shut down access to their V1 API which did not require an API key. For this reason V2 has now been implemented. An API key is now required. The user will be prompted for this key when running populate.py or profile.yaml can be manually edited.

More details on obtaining a key will be posted to the readme soon.

Depending on performance and usage caps, Google's API may be abandoned in the future in favor of AT&T's API.