Fix keyword detection regression in Mic.passiveListen()
Prior to commit 6bb87e20423e41f1bb79e6af285fe4d49cf1359a (Handle all Google Speech options, Pull Request #148), passive listen checked if the transcribed phrase contained the keyword. Since then, it only detects if the transcribed phrase was *exactly* the keyword. When using Pocketsphinx, it often happens that additional words are transcribed (like "JASPER OF"), so that saying the keyword sometimes did not result in an active listen call. With this commmit, this regression will be fixed and the old behaviour restored.