Fix Weather testcase
The only change here is that the test profile location has been changed from "08544" (US zipcode of Princeton, NJ) to "Cape Town" (this was chosen because of high code coverage). It looks like it was broken anyway and just passed because the weather module did not fail if the city was not found. If I type http://rss.wunderground.com/auto/rss_full/08544 into my webbrowser, it says "City not found".

If you now have an invalid location name in your profile, the user will be informed that something is fishy.