Add workaround for buggy pocketsphinx import
Certain versions of pocketsphinx seem to have a bug which causes a
ValueError ("PyCapsule_GetPointer called with invalid PyCapsule object")
when trying to import pocketsphinx for the first time. It does work when
importing a second time. Thus, we just work around this by importing it
again if the first import raised a ValueError.

See http://sourceforge.net/p/cmusphinx/bugs/284/ for details.

This should (partly) resolve #391.